requirements.txt
rembg
Pillow
onnxruntime